refactor: move initialization of config, db, ws, server to main.go
